pyproject.toml
[tool.poetry]
name = "ada-mata-mle"
version = "0.1.0"
description = "Edge-Optimized Bottle Cap Detector (YOLOv8 + OpenVINO)"
authors = ["wikan1602 <wikanpriambudi@gmail.com>"]
readme = "README.md"
# PENTING: Ini agar folder src/bsort terbaca sebagai package
packages = [{include = "bsort", from = "src"}]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# --- CLI ENTRY POINT ---
[tool.poetry.scripts]
bsort = "bsort.main:app"

# --- SOURCE PYTORCH CPU (Agar Docker Image Kecil) ---
[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

# --- MAIN DEPENDENCIES ---
[tool.poetry.dependencies]
python = "^3.11"

# 1. Core Logic
ultralytics = "^8.3"         # Versi terbaru untuk stabilitas export
opencv-python-headless = "*" # Versi headless (tanpa GUI) untuk server/docker
pyyaml = "*"

# 2. CLI Interface
# KITA KUNCI KE VERSI BARU AGAR TIDAK CRASH (Error make_metavar)
typer = { version = "0.9.0", extras = ["all"] }
click = "8.1.7"

# 3. Experiment Tracking
wandb = "*"

# 4. Edge Deployment & Export
# ONNX wajib ada sebagai format perantara sebelum ke OpenVINO
onnx = "*"
openvino = ">=2024.0.0,<2025.0.0"
nncf = ">=2.10.0,<2.13.0"

# 5. Deep Learning Framework (CPU Only)
# Kita kunci versi spesifik agar reproducible
torch = { version = "2.4.1", source = "pytorch-cpu" }
torchvision = { version = "0.19.1", source = "pytorch-cpu" }

# --- DEVELOPMENT & TESTING ---
[tool.poetry.group.dev.dependencies]
pytest = "*"
pylint = "*"
black = "*"
isort = "*"
jupyter = "*"
ipykernel = "*" # Tambahan agar kernel notebook terdeteksi